Consolidate actions
Will split in a later commit
